avm2: Add convenience method for natively-dispatched events.
